PD #82500: Add filter to avoid OS power off by loading jitter
When battery is over-discharging and boot up with extern power(usually battery
capacity is 0%). If CPU loading and battery backlight level are high, extern power
is not enough for system use and battery will discharge now. To protect battery,
if we detect battery is under discharging state over 10 seconds, then we will
force OS to power off.