PD  #90113: dirvers: fix kernel detect task hungup issue.
In the Kernel thread, msleep is uninterruptible wait. The status of the task is "D".
And the Kernel thread is uninterruptible sleep after it's created but not run.
The Kernel hung_task detect checks all "D" status task for several seconds. If no any
context switch changed, the task will be reported as the "hung task".

This patch changs msleep() to msleep_interruptible(), and wakes up kernel thread after
it's created.


Former-commit-id: 8b8bd9321634a64e61fd77eec6ee622f086467a3